image-browser-app/requirements.txt
streamlit
Pillow
opencv-python
numpy